+++ test/ecosystem.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+    apps : [{
+      name   : "worker",
+      script : "./test.js",
+      args   : "limit"
+    }]
+  }
+