+++ test/test.js
@@ -1,9 +1,13 @@
-var MB = '';
+let MBout = '';
+let MBerr = '';
 
-for (var i = 0; i < 1024 * 1024; ++i)
-  MB += '1';
+for (let i = 0; i < 1024 * 1024; ++i)
+  MBout += 'o';
+
+for (let i = 0; i < 1024 * 1024; ++i)
+  MBerr += 'e';
 
 setInterval(function() {
-  process.stdout.write(MB);
-  process.stderr.write(MB);
+  process.stdout.write(MBout);
+  process.stderr.write(MBerr);
 }, 1000);